refactor: extract manage-paths reconcile into shared module
